UniqueDigits_test->UniqueDigits_sub_test, UnqiueDigit_test -> UniqueDigits_test
