+++ src/test/java/UniqueDigits_test.java
@@ -4,37 +4,27 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import org.junit.jupiter.api.Test;
 
 public class UniqueDigits_test {
-    @Test
-    public void UniqueDigitsSubTest() {
-        // 문자열 내 중복되는 문자가 하나라도 존재한다면 유효하지 않은 문자열
-        String number = "12345";
-        assertTrue(BaseballGame.UniqueDigits_sub(0, number));
-        assertTrue(BaseballGame.UniqueDigits_sub(1, number));
-        assertTrue(BaseballGame.UniqueDigits_sub(2, number));
-        assertTrue(BaseballGame.UniqueDigits_sub(3, number));
-        assertTrue(BaseballGame.UniqueDigits_sub(4, number));
-    }
 
-    @Test
-    public void UniqueDigitsSubTest2() {
-        // 문자열 내 중복되는 문자가 하나라도 존재한다면 유효하지 않은 문자열
-        String number = "11234";
-        assertFalse(BaseballGame.UniqueDigits_sub(0, number));
-        assertTrue(BaseballGame.UniqueDigits_sub(1, number));
-        assertTrue(BaseballGame.UniqueDigits_sub(2, number));
-        assertTrue(BaseballGame.UniqueDigits_sub(3, number));
-        assertTrue(BaseballGame.UniqueDigits_sub(4, number)); // 마지막 문자는 중복이 없으므로 true
-    }
+        @Test
+        public void testUniqueDigits() {
+            assertTrue(BaseballGame.UniqueDigits("123"));
+            assertTrue(BaseballGame.UniqueDigits("456"));
+            assertTrue(BaseballGame.UniqueDigits("789"));
+            assertFalse(BaseballGame.UniqueDigits("111"));
+            assertFalse(BaseballGame.UniqueDigits("222"));
+            assertFalse(BaseballGame.UniqueDigits("333"));
+            assertFalse(BaseballGame.UniqueDigits("444"));
+            assertFalse(BaseballGame.UniqueDigits("555"));
+            assertFalse(BaseballGame.UniqueDigits("666"));
+            assertFalse(BaseballGame.UniqueDigits("777"));
+            assertFalse(BaseballGame.UniqueDigits("888"));
+            assertFalse(BaseballGame.UniqueDigits("999"));
+            assertFalse(BaseballGame.UniqueDigits("000"));
+            assertFalse(BaseballGame.UniqueDigits("001"));
+            assertFalse(BaseballGame.UniqueDigits("010"));
+            assertFalse(BaseballGame.UniqueDigits("100"));
+        }
 
-    @Test
-    public void UniqueDigitsSubTest3() {
-        // 문자열 내 중복되는 문자가 하나라도 존재한다면 유효하지 않은 문자열
-        String number = "12234";
-        assertTrue(BaseballGame.UniqueDigits_sub(0, number));
-        assertFalse(BaseballGame.UniqueDigits_sub(1, number));
-        assertTrue(BaseballGame.UniqueDigits_sub(2, number));
-        assertTrue(BaseballGame.UniqueDigits_sub(3, number));
-        assertTrue(BaseballGame.UniqueDigits_sub(4, number)); // 마지막 문자는 중복이 없으므로 true
-    }
+
 
 }
